requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
moviepy==1.0.3
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==3.0.1